fix: 補上 messageValues 缺漏的 type cast 並修正 SUM 回傳 BigInt 問題
- viewer-message.repository.ts：messageValues 補上 ::text cast（fingerprint、viewerId、channelId、messageText、messageType、badges、emotesUsed），避免 null 值在 PostgreSQL VALUES CTE 型別推斷失敗
- viewer.service.ts：SUM watchSeconds 補上 ::float8 cast，防止 bigint 在 JSON 序列化時出錯

+++ backend/src/modules/viewer/viewer.service.ts
@@ -802,7 +802,7 @@ async function buildFollowedChannelsFromSource(viewerId: string): Promise<Follow
       WHERE "viewerId" = ${viewerId}
     ),
     daily_watch AS (
-      SELECT "channelId", SUM("watchSeconds") / 60 AS "dailyWatchMin"
+      SELECT "channelId", SUM("watchSeconds")::float8 / 60 AS "dailyWatchMin"
       FROM viewer_channel_daily_stats
       WHERE "viewerId" = ${viewerId}
       GROUP BY "channelId"
@@ -908,7 +908,7 @@ export async function syncSummaryStatsFromLifetime(viewerId: string): Promise<vo
             0
           ),
           COALESCE(
-            (SELECT SUM(d."watchSeconds") / 60
+            (SELECT SUM(d."watchSeconds")::float8 / 60
              FROM viewer_channel_daily_stats d
              WHERE d."viewerId" = viewer_channel_summary."viewerId"
                AND d."channelId" = viewer_channel_summary."channelId"),
